backend/requirements.txt
base36==0.1.1
Flask-Limiter==2.0.4
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
# Flask-Sockets==0.2.1
git+git://github.com/ebrensi/flask-sockets.git
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask==2.0.2
gevent==21.12.0
gunicorn==20.1.0
msgpack==1.0.3
polyline==1.4.0
psycopg2==2.9.2
pymongo==4.0.1
python-dateutil==2.8.2
redis==4.0.2
requests==2.26.0
stravalib==0.10.4
whitenoise==5.3.0
wsaccel==0.6.3

